Fixed bugs in the client scores section on the survey page by removing extra spaces or inconsistencies in the data such as different capitalization.
Fixed bug where client evaluation was not showing up. The issue was related to how it does checking for the non-empty questions. In the previous code, we were appending question numbers to the @not_empty_questions array, whereas in the updated code, we did not update this array. To fix this, we needed to modify the code to update the @not_empty_questions array based on the presence of question values in the filtered @cliSurvey data.

This should fix the issue, and the client evaluation data should be displayed on the page for the relevant questions.